fixed error with missed access_log table on tests where clickhouse not required

+++ helpers/clickhouse.py
@@ -91,6 +91,11 @@ class ClickHouseFinder(dmesg.BaseTempestaLogger):
         """
         Delete all log records
         """
+        response = self._clickhouse_client.query("exists table access_log")
+
+        if not response.result_rows[0][0]:
+            return
+
         self._clickhouse_client.command("delete from access_log where true")
 
     def access_log_records_count(self) -> int:
